Added unit tests for InitializeK8sAuth class
Also moved most of the logic for unit testing
initializer classes to a helper file to reduce
duplication.
